fix HydrometTools not showing left-axis units
On both the Check->VolumeInRange and Check->Monthly the left-axis units were not visible for heii qu. I think the scale of the y-axis labels pushed the units to not visible, and I think making the right-axis visible pushed the units to not visible. Calling tchart.Draw() seems to redraw the graph and everything is visible.